// lib/plant_tracker/model/locationOut.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Plant Tracker API
 * API for interacting with the Plant Tracker application
 * OpenAPI spec version: 1.0.0
 */
import type { UserSchema } from "./userSchema";

export interface LocationOut {
  id: string;
  name: string;
  user: UserSchema;
}